Change Simple Interest calculation to use float

+++ src/CH12.java
@@ -12,7 +12,7 @@ public class CH12
         float t = sc.nextFloat();
         System.out.print("Enter the Rate of Interest in percentage : ");
         float r = sc.nextFloat();
-        double si = (p*t*r)/100;
+        float si = (p*t*r)/100;
         System.out.print("\nThe Simple Interest is "+si);
         sc.close();
     }
